Add preprocessing config work-in-progress
