+++ ml/bluenot.py
@@ -33,7 +33,7 @@ import pathlib
 import subprocess
 import time
 from argparse import ArgumentParser
-from typing import Dict, List, Tuple, Union, Sequence
+from typing import Dict, List, Tuple, Union
 
 import keras
 import numpy as np
@@ -47,8 +47,6 @@ from blueno import (
     utils,
     elasticsearch,
     io,
-    preprocessing,
-    transforms,
 )
 
 os.environ['CUDA_VISIBLE_DEVICES'] = ''
@@ -115,90 +113,6 @@ def to_arrays(data: Dict[str, np.ndarray],
     return np.stack(X_list), np.stack(y_list), np.array(remaining_ids)
 
 
-RAW_ARRAYS = {}
-
-
-def preprocess_data(data_dir: str,
-                    labels_path: str,
-                    height_offset: int,
-                    mip_thickness: int,
-                    pixel_value_range: Sequence) -> Tuple[dict, pd.DataFrame]:
-    """
-    Preprocesses data by loading raw numpy data, cleaning, filtering,
-    and transforming the arrays and labels.
-
-    It is recommended that you use your own preprocessing callable
-    if you want more configurability.
-
-    :param data_dir: path to a directory containing .npz files
-    :param labels_path: path a directory containing the 2 metadata CSVs
-    :param height_offset: the offset from the top of the 3D image to crop from
-    :param mip_thickness: the thickness of a single mip
-    :param pixel_value_range: the range of pixel values to crop to
-    :return:
-    """
-    global RAW_ARRAYS
-    if RAW_ARRAYS:
-        logging.info('loading cached arrays')
-        raw_arrays = RAW_ARRAYS
-    else:
-        logging.info('loading raw arrays from disk')
-        raw_arrays = io.load_compressed_arrays(data_dir)
-        RAW_ARRAYS = raw_arrays
-
-    raw_labels = io.load_labels(labels_path)
-
-    cleaned_arrays, cleaned_labels = preprocessing.clean_data(raw_arrays,
-                                                              raw_labels)
-
-    def filter_data(arrays: dict, labels: pd.DataFrame):
-        filtered_arrays = {id_: arr for id_, arr in arrays.items()
-                           if arr.shape[0] != 1}  # Remove the bad array
-        filtered_labels = labels.loc[filtered_arrays.keys()]
-        assert len(filtered_arrays) == len(filtered_labels)
-        return filtered_arrays, filtered_labels
-
-    filtered_arrays, filtered_labels = filter_data(cleaned_arrays,
-                                                   cleaned_labels)
-
-    def process_data(arrays: dict,
-                     labels: pd.DataFrame,
-                     height_offset: int,
-                     mip_thickness: int,
-                     pixel_value_range: Sequence):
-        processed_arrays = {}
-        for id_, arr in arrays.items():
-            try:
-                arr = transforms.crop(arr,
-                                      (3 * mip_thickness, 200, 200),
-                                      height_offset=height_offset)
-                arr = np.stack([arr[:mip_thickness],
-                                arr[mip_thickness: 2 * mip_thickness],
-                                arr[2 * mip_thickness: 3 * mip_thickness]])
-                arr = transforms.bound_pixels(arr,
-                                              pixel_value_range[0],
-                                              pixel_value_range[1])
-                arr = arr.max(axis=1)
-                arr = arr.transpose((1, 2, 0))
-                processed_arrays[id_] = arr
-            except AssertionError:
-                print(f'patient id {id_} could not be processed,'
-                      f' has input shape {arr.shape}')
-        processed_labels = labels.loc[
-            processed_arrays.keys()]  # Filter, if needed
-        assert len(processed_arrays) == len(processed_labels)
-        return processed_arrays, processed_labels
-
-    processed_arrays, processed_labels = process_data(
-        filtered_arrays,
-        filtered_labels,
-        height_offset=height_offset,
-        mip_thickness=mip_thickness,
-        pixel_value_range=pixel_value_range
-    )
-    return processed_arrays, processed_labels
-
-
 def prepare_data(params: blueno.ParamConfig) -> Tuple[np.ndarray,
                                                       np.ndarray,
                                                       np.ndarray,
@@ -215,11 +129,25 @@ def prepare_data(params: blueno.ParamConfig) -> Tuple[np.ndarray,
     logging.info(f'using params:\n{params}')
     # Load the arrays and labels
     data_params = params.data
-    array_dict = io.load_arrays(data_params.data_dir)
-    index_col = data_params.index_col
+    if isinstance(data_params, blueno.LukePipelineConfig):
+        array_dict, labels_df = data_params.pipeline_callable(
+            data_params.data_dir,
+            data_params.labels_path,
+            data_params.height_offset,
+            data_params.mip_thickness,
+            data_params.pixel_value_range,
+        )
+    elif isinstance(data_params, blueno.DataConfig):
+        array_dict = io.load_arrays(data_params.data_dir)
+        index_col = data_params.index_col
+        labels_df = pd.read_csv(data_params.labels_path,
+                                index_col=index_col)
+    else:
+        raise ValueError('params.data must be a subclass of DataConfig')
+
     label_col = data_params.label_col
-    label_series = pd.read_csv(data_params.labels_path,
-                               index_col=index_col)[label_col]
+    label_series = labels_df[label_col]
+
     # Convert to numpy arrays
     x, y, patient_ids = to_arrays(array_dict, label_series)
 
@@ -430,6 +358,7 @@ def hyperoptimize(hyperparams: Union[blueno.ParamGrid,
             params = blueno.ParamConfig(**params)
         # This is where we'd run preprocessing. To run in a reasonable amount
         # of time, the raw data must be cached in-memory.
+
         x_train, x_valid, y_train, y_valid, id_train, id_valid = prepare_data(
             params)
 
@@ -516,7 +445,7 @@ if __name__ == '__main__':
     elif isinstance(user_config.PARAM_GRID, dict):
         logging.warning('creating param grid from dictionary, it is'
                         'recommended that you define your config'
-                        'with ParamGrid')
+                        'with ParamConfig')
         param_grid = blueno.ParamGrid(**user_config.PARAM_GRID)
     else:
         raise ValueError('user_config.PARAM_GRID must be a ParamGrid,'
